Added the capability to construct a setcover from  a coverage report like is expected by the R CE calculator, i.e.
NameCD	Requirements
1	1
1	45
1	23
2	4
3	1
.	.
.	.
.	.

Use the constructSetCoverFromCoverageAndTime method and pass the 
coverage and time file names as parameters.

The major benefit to using this format is that it allows you to
keep the index numbers of the tests and requirements.  The matrix
format does not save indexes, and renumbers the tests when they are
imported.

In the future, it may be a good (great?) idea to save the name
of the test in the coverage file.  

Also it would be great to be able to trace the source of the 
requirement.